refactor(modeloanimal): consolidar lógica, registrar inoculación y migrar a BigDecimal

+++ backend/src/main/java/com/utn/magtea/modeloanimal/ModeloAnimalPoolAporteDTO.java
@@ -1,8 +1,10 @@
 package com.utn.magtea.modeloanimal;
 
+import java.math.BigDecimal;
+
 public record ModeloAnimalPoolAporteDTO(
         Long poolTuboId,
         String posicion,
-        Double cantidadConsumida,
+        BigDecimal cantidadConsumida,
         Integer dia
 ) {}
